Match Piano/Marimba hands to direction on keyboard
Left click ( and A ) plays the left bongo, Right click ( and D )
plays the right bongo. It is therefore logical that the keyboard
keys 1-5, and Q-T play the left Piano and Marimba keys respectively,
and that the keyboard keys 6-0 and Y-P play the right Piano and Marimba
keys respectively.

This patch fixes this visual bug and soothes my OCD.