Allow any combination of simultaneous sounds
Remove the check preventing you from using the "same paw" simultaneously.
There's no particular reason the cat should be allowed to play keyboard 2 and marimba Q at the same time (both are left paw), but not be allowed to play keyboard 1 and marimba Q at the same time. This is especially silly with keyboard 1 and bongo A, which are displayed as different paws, but not allowed to be played simultaneously. This behavior is confusing and unnecessary. Remove this rule so you can play whatever you want.
This also fixes issue #145.
This change has the side effect of letting you play the bongos very fast by using the keyboard and mouse, but that's fine imo.