Ode to Joy melody is misattributed (Fixes #168) (#170)
The Ode to Joy melody was written by Ludwig van Beethoven for his 9th Symphony. Friedrich Schiller wrote the poem Ode to Joy which is part of it is used as the lyrics to the final movement in Beethoven symphony. In the README.md it attributes the melody of Ode to Joy to Friedrich Schiller which is wrong, the melody was written by Beethoven.